% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BytesToDouble}
\alias{BytesToDouble}
\title{Converting \code{sizeof(double)} bytes to a double value}
\usage{
BytesToDouble(x)
}
\arguments{
\item{x}{A vector of bytes (unsigned chars in C++)}
}
\value{
Corresponding \code{double} typed value for a given vector of bytes
}
\description{
This function converts \code{sizeof(double)} bytes to a double typed value
}
\examples{
print(BytesToDouble(DoubleToBytes(56.43)))
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
\seealso{
DoubleVectorToBytes

DoubleToBytes

ByteVectorToDoubles
}

