% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ByteVectorToDoubles}
\alias{ByteVectorToDoubles}
\title{Converting p * \code{sizeof(double)} bytes to a vector of p double values}
\usage{
ByteVectorToDoubles(b)
}
\arguments{
\item{b}{A vector of bytes (unsigned chars in C++)}
}
\value{
Corresponding vector of \code{double} typed values for a given vector of bytes
}
\description{
This function converts a byte vector to a vector of doubles
}
\examples{
a <- DoubleVectorToBytes(c(56.54, 89.7666, 98.565))
b <- ByteVectorToDoubles(a)
print(b)
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
\seealso{
DoubleVectorToBytes

BytesToDouble

ByteVectorToDoubles
}

