% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DoubleToBytes}
\alias{DoubleToBytes}
\title{Byte representation of a \code{double} typed variable}
\usage{
DoubleToBytes(x)
}
\arguments{
\item{x}{A double typed value}
}
\value{
A vector of byte values with the length of \code{sizeof(double)} for a given \code{double} typed value
}
\description{
This function returns a vector of byte values with the length of \code{sizeof(double)} for a given \code{double} typed value
}
\examples{
print(DoubleToBytes(56.43))
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
\seealso{
DoubleVectorToBytes

BytesToDouble

ByteVectorToDoubles
}

