% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DoubleVectorToBytes}
\alias{DoubleVectorToBytes}
\title{Byte representation of a vector of \code{double} typed variables}
\usage{
DoubleVectorToBytes(d)
}
\arguments{
\item{d}{A vector of double typed values}
}
\value{
returns a vector of byte values for a given vector of \code{double} typed values
}
\description{
This function returns a vector of byte values for a given vector of \code{double} typed values
}
\examples{
print(DoubleVectorToBytes(c(56.54, 89.7666, 98.565)))
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
\seealso{
DoubleToBytes

BytesToDouble

ByteVectorToDoubles
}

