% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaxDouble}
\alias{MaxDouble}
\title{Maximum value of a \code{double} typed variable}
\usage{
MaxDouble()
}
\value{
Returns maximum value of a \code{double} typed variable in C++ compiler
}
\description{
Maximum value of a \code{double} typed variable
}
\examples{
print(MaxDouble())
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}

