% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SizeOfDouble}
\alias{SizeOfDouble}
\title{Byte-length of a \code{double} typed variable}
\usage{
SizeOfDouble()
}
\value{
Returns the byte-length of a \code{double} typed variable in computer memory
}
\description{
Byte-length of a \code{double} typed variable in computer memory
}
\examples{
print(SizeOfDouble())
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}

