% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SizeOfInt}
\alias{SizeOfInt}
\title{Byte-length of a \code{int} typed variable}
\usage{
SizeOfInt()
}
\value{
Returns the byte-length of a \code{int} typed variable in computer memory
}
\description{
Byte-length of a \code{int} typed variable in computer memory
}
\examples{
print(SizeOfInt())
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}

