% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SizeOfLong}
\alias{SizeOfLong}
\title{Byte-length of a \code{long} typed variable}
\usage{
SizeOfLong()
}
\value{
Returns the byte-length of a \code{long} typed variable in computer memory
}
\description{
Byte-length of a \code{long} typed variable in computer memory
}
\examples{
print(SizeOfLong())
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}

