% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TwoPointCrossOver}
\alias{TwoPointCrossOver}
\title{Two Point Crossover operation on the two vectors of bytes}
\usage{
TwoPointCrossOver(bytes1, bytes2, cutpoint1, cutpoint2)
}
\arguments{
\item{bytes1}{A vector of bytes of the first parent}

\item{bytes2}{A vector of bytes of the second parent}

\item{cutpoint1}{First cut-point for the single point crossing-over}

\item{cutpoint2}{Second cut-point for the single point crossing-over}
}
\value{
List of two byte vectors of offspring
}
\description{
This function is a C++ wrapper for crossing-over of two byte vectors of candidate solutions
}
\examples{
b1 <- DoubleVectorToBytes(c(56.54, 89.7666, 98.565))
b2 <- DoubleVectorToBytes(c(79.76, 56.4443, 34.22121))
cutpoints <- sort(sample(1:(length(b1)*SizeOfDouble()), 2, replace = FALSE))
result <- TwoPointCrossOver(b1,b2, cutpoints[1], cutpoints[2])
print(ByteVectorToDoubles(result[[1]]))
print(ByteVectorToDoubles(result[[2]]))
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
\seealso{
OnePointCrossOver

OnePointCrossOverOnDoublesUsingBytes

UniformCrossOverOnDoublesUsingBytes
}

