\name{mcga-package}
\alias{mcga-package}
\alias{mcga-package}
\docType{package}
\title{
Machine Coded Genetic Algorithms for the Real Value Optimization Problems.
}
\description{
mcga package uses machine coded chromosomes for real value optimization problems. 
Chromosomes are not encoded/decoded, the byte representation of the double values are used for crossing-over and mutation operations. 
This makes mcga extremely fast even the number of parameters is high. 
Values of the parameters are bounded with the machine double type.

In version 1.1, some performance improvements are done. mcga function now returns the entire population and the best solution vector.
}
\details{
\tabular{ll}{
Package: \tab mcga\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2010-12-22\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Mehmet Hakan Satman

Maintainer: Mehmet Hakan Satman <mhsatman@istanbul.edu.tr>
}
